new function to pull SBD telemetry directly (#9)
New function pullSWIFTtelemetry, which uses Matlab's websave to query the SWIFT server for a list of SWIFT IDs over a certain time, then read everything with compileSWIFT_SBDtelemetry (and readSWIFT_SBD) and make some overview plots.  This is intended for situational awareness during a project/mission with lots of SWIFTs... including reporting battery voltage.  It avoids having to manually query the server with the web form.